requirements.txt
requests==2.25.1
ConfigArgParse == 1.3
discord.py==1.6.0
windows-curses==2.2.0; platform_system == "Windows"